Changing how scene loading works
Scene loading now works additively, meaning that assets from another scene will be added on top of the existing scene, instead of two discrete scenes that must be swapped between. Scene Unloading is being added, but does not work correctly in this version.
Additionally creating an extra backup scene which is a copy of TestSceneDeux in case something got borked in transit and additive loading didn't work correctly.