Continued modifications for scene swapping
Created some facts for the game engine to use for basic dialogue and scene change testing, as well as edits within the .yarn file  towards this end with a simple dialogue tree. Also added a debug message to confirm what scenes are being unloaded, a new scene to confirm if facts are being remembered/stored (they are not, at the moment), and moved the capsules in TestSceneDeux so it is easier to tell when a scene change has occurred.